requirements.txt
asgiref==3.5.2
Django==4.0.6
django-autoslug==1.9.8
Pillow==9.2.0
sqlparse==0.4.2
tzdata==2022.1
whitenoise==6.2.0
